tests/smoke/test_auth.py
"""Smoke tests for authentication functionality.

Tests authentication flows against deployed environment.
"""

import os

import pytest
import requests


@pytest.fixture
def base_url() -> str:
    """Get base URL from environment or use default."""
    return os.getenv(
        "SMOKE_TEST_URL",
        "http://sakese-Publi-5SDe3QrKne55-1360562030.us-west-2.elb.amazonaws.com",
    )


@pytest.fixture
def cognito_config() -> dict[str, str]:
    """Get Cognito configuration from environment."""
    return {
        "region": os.getenv("AWS_REGION", "us-west-2"),
        "user_pool_id": os.getenv("COGNITO_USER_POOL_ID", ""),
        "client_id": os.getenv("COGNITO_CLIENT_ID", ""),
    }


@pytest.mark.smoke
class TestAuthenticationSmoke:
    """Smoke tests for authentication functionality."""

    def test_app_loads_without_auth(self, base_url: str) -> None:
        """Test that application loads without authentication."""
        response = requests.get(base_url, timeout=30, allow_redirects=True)

        # Application should load (200) or redirect (3xx)
        assert response.status_code in [200, 301, 302, 303, 307, 308]

    def test_cognito_config_available(self, cognito_config: dict[str, str]) -> None:
        """Test that Cognito configuration is available."""
        # In deployed environment, these should be set
        # For local testing, they may be empty
        assert "region" in cognito_config
        assert "user_pool_id" in cognito_config
        assert "client_id" in cognito_config

        # Region should always have a value
        assert cognito_config["region"]
        assert len(cognito_config["region"]) > 0

    def test_auth_page_accessible(self, base_url: str) -> None:
        """Test that authentication page is accessible."""
        # Streamlit apps serve all content through main URL
        response = requests.get(base_url, timeout=30)

        # Should get a valid response
        assert response.status_code == 200

        # Response should be HTML
        assert "text/html" in response.headers.get("Content-Type", "")

    def test_invalid_credentials_pattern(self) -> None:
        """Test invalid credentials error pattern validation."""
        # This tests the error handling pattern without actually calling Cognito
        # In a real scenario, invalid credentials should raise NotAuthorizedException

        error_patterns = {
            "NotAuthorizedException": "Invalid credentials or user not found",
            "UserNotFoundException": "User does not exist",
            "UserNotConfirmedException": "User email not verified",
            "PasswordResetRequiredException": "Password reset required",
            "TooManyRequestsException": "Too many failed attempts",
        }

        for error_code, description in error_patterns.items():
            # Validate that error codes follow AWS Cognito patterns
            assert error_code.endswith("Exception")
            assert len(description) > 0
            assert isinstance(description, str)

    def test_session_state_pattern(self) -> None:
        """Test session state management patterns."""
        # Validate session state keys that should be used
        expected_session_keys = [
            "authenticated",
            "user_id",
            "username",
            "email",
            "id_token",
            "access_token",
            "refresh_token",
        ]

        for key in expected_session_keys:
            # Validate key naming conventions
            assert isinstance(key, str)
            assert len(key) > 0
            assert "_" in key or key.islower()  # snake_case or lowercase

    def test_auth_flow_sequence(self) -> None:
        """Test authentication flow sequence validation."""
        # Document the expected authentication flow
        auth_flow_steps = [
            "1. User accesses application",
            "2. Application checks session state",
            "3. If not authenticated, show login/signup",
            "4. User enters credentials",
            "5. Application calls Cognito InitiateAuth",
            "6. Cognito returns tokens (access, id, refresh)",
            "7. Application stores tokens in session",
            "8. User is redirected to main page",
        ]

        # Validate flow documentation
        assert len(auth_flow_steps) == 8
        for step in auth_flow_steps:
            assert step.startswith(str(auth_flow_steps.index(step) + 1))
            assert len(step) > 10  # Each step should have meaningful description

    def test_token_validation_requirements(self) -> None:
        """Test JWT token validation requirements."""
        # Document required JWT claims for Cognito tokens
        required_id_token_claims = [
            "sub",  # User ID
            "email",  # User email
            "email_verified",  # Email verification status
            "cognito:username",  # Username
            "exp",  # Expiration time
            "iat",  # Issued at time
            "token_use",  # Should be "id"
        ]

        required_access_token_claims = [
            "sub",  # User ID
            "exp",  # Expiration time
            "iat",  # Issued at time
            "token_use",  # Should be "access"
            "scope",  # Token scope
            "client_id",  # Cognito client ID
        ]

        # Validate claim lists
        assert len(required_id_token_claims) >= 5
        assert len(required_access_token_claims) >= 5
        assert "sub" in required_id_token_claims
        assert "sub" in required_access_token_claims
        assert "exp" in required_id_token_claims
        assert "exp" in required_access_token_claims


@pytest.mark.smoke
class TestAuthenticationSecurity:
    """Smoke tests for authentication security."""

    def test_https_enforcement_pattern(self, base_url: str) -> None:
        """Test HTTPS enforcement pattern (for production)."""
        # In production, should use HTTPS
        # In dev/staging with ALB, may use HTTP with ALB terminating SSL

        if base_url.startswith("https://"):
            # If HTTPS, verify it's accessible
            response = requests.get(base_url, timeout=30, verify=True)
            assert response.status_code in [200, 301, 302, 303, 307, 308]
        elif base_url.startswith("http://"):
            # If HTTP (dev environment), verify it's accessible
            response = requests.get(base_url, timeout=30)
            assert response.status_code in [200, 301, 302, 303, 307, 308]

    def test_password_policy_requirements(self) -> None:
        """Test password policy requirements validation."""
        # Document Cognito password requirements
        password_requirements = {
            "min_length": 8,
            "require_uppercase": True,
            "require_lowercase": True,
            "require_numbers": True,
            "require_special_chars": True,
        }

        # Validate requirements
        assert password_requirements["min_length"] >= 8
        assert password_requirements["require_uppercase"] is True
        assert password_requirements["require_numbers"] is True

    def test_session_timeout_pattern(self) -> None:
        """Test session timeout configuration pattern."""
        # Document expected session timeout behavior
        timeout_config = {
            "id_token_validity": 3600,  # 1 hour
            "access_token_validity": 3600,  # 1 hour
            "refresh_token_validity": 86400 * 30,  # 30 days
        }

        # Validate timeout values are reasonable
        assert timeout_config["id_token_validity"] > 0
        assert timeout_config["access_token_validity"] > 0
        assert timeout_config["refresh_token_validity"] > timeout_config["id_token_validity"]

    def test_rate_limiting_pattern(self) -> None:
        """Test rate limiting pattern for authentication."""
        # Document expected rate limiting behavior
        rate_limits = {
            "login_attempts": 5,  # Max failed attempts before lockout
            "lockout_duration": 300,  # 5 minutes
            "signup_rate": 10,  # Max signups per hour per IP
        }

        # Validate rate limit values
        assert rate_limits["login_attempts"] > 0
        assert rate_limits["lockout_duration"] > 0
        assert rate_limits["signup_rate"] > 0
